# ENH: Use AVX for float32 implementation of np.sin & np.cos