--- a/numpy/core/src/umath/cpuid.c
+++ b/numpy/core/src/umath/cpuid.c
@@ -48,6 +48,25 @@ int os_avx512_support(void)
 #endif
 }
 
+static NPY_INLINE
+int cpu_supports_fma(void)
+{
+    unsigned int feature = 0x01;
+    unsigned int a, b, c, d;
+#ifdef __x86_64__
+    __asm__ volatile (
+        "cpuid"				"\n\t"
+	: "=a" (a), "=b" (b), "=c" (c), "=d" (d)
+	: "a" (feature));
+    /*
+     * FMA is the 12th bit of ECX
+     */
+    return (c >> 12) & 1;
+#else
+    return 0;
+#endif
+}
+
 /*
  * Primitive cpu feature detect function
  * Currently only supports checking for avx on gcc compatible compilers.
@@ -62,6 +81,9 @@ npy_cpu_supports(const char * feature)
 #else
         return 0;
 #endif
+    }
+    else if (strcmp(feature, "fma") == 0) {
+        return cpu_supports_fma() && __builtin_cpu_supports("avx2") && os_avx_support();
     }
     else if (strcmp(feature, "avx2") == 0) {
         return __builtin_cpu_supports("avx2") && os_avx_support();