--- a/numpy/core/src/umath/loops.c.src
+++ b/numpy/core/src/umath/loops.c.src
@@ -1645,17 +1645,17 @@ FLOAT_@func@_@isa@(char **args, npy_intp *dimensions, npy_intp *steps, void *NPY
 
 /**begin repeat1
  *  #func = cos, sin#
+ *  #enum = npy_compute_cos, npy_compute_sin#
  *  #scalarf = npy_cosf, npy_sinf#
  */
 
 NPY_NO_EXPORT NPY_GCC_OPT_3 void
 FLOAT_@func@_@isa@(char **args, npy_intp *dimensions, npy_intp *steps, void *NPY_UNUSED(data))
 {
-    char str[] = "@func@";
-    if (!run_unary_@isa@_sincos_FLOAT(args, dimensions, steps, str)) {
+    if (!run_unary_@isa@_sincos_FLOAT(args, dimensions, steps, @enum@)) {
         UNARY_LOOP {
 #if defined @CHK@ && defined NPY_HAVE_SSE2_INTRINSICS
-            @ISA@_sincos_FLOAT((npy_float *)op1, (npy_float *)ip1, 1, steps[0], str);
+            @ISA@_sincos_FLOAT((npy_float *)op1, (npy_float *)ip1, 1, steps[0], @enum@);
 #else
 	        const npy_float in1 = *(npy_float *)ip1;
 	        *(npy_float *)op1 = @scalarf@(in1);