… numpy/core/code_generators/generate_umath.py
… numpy/core/tests/test_umath.py