@r-devulap 2019-04-19
I realize the first 6 commits are exactly same as in this PR: https://github.com/numpy/numpy/pull/13134. The only commit that is new is the last one which implements sine and cosine (but it relies on the previous commits).