@eric-wieser 2019-04-20
For reviewers - if the other PR is merged, you can switch the base of this PR between <anything> and back to `master`, and it will recompute the diff and throw out all the merged commits.