@r-devulap 2019-04-23
As mentioned in the commit message, the performance gains are leveraged only if the input arguments to sine/cosine are between [-71476.0625f, 71476.0625f] for cosine and [-117435.992f,
117435.992f] for sine.  I measured 3 different benchmarks.

(1) For micro-benchmarks: 
Array size = 10000, Command = "%timeit np.cos([myarr])"

| Function Name | NumPy 1.16 | AVX2   | AVX512  | AVX2 speed up   | AVX512  speed up |
| ------------- | ----------- | -------- | --------- | ---------- | ---------- |
| np.cos        | 1.5174s     | 0.1553s | 0.06634s | 9.77x     | 22.87x    |
| np.sin        | 1.4738s     | 0.1517s | 0.06406s | 9.71x     | 23.00x    |

(2) Package ai.cs provides an API to transform spherical coordinates to
cartesean system:
Array size = 10000, Command = "%timeit ai.cs.sp2cart(r,theta,phi)"

| Function Name | NumPy 1.16 | AVX2   | AVX512  | AVX2 speed up   | AVX512  speed up |
| ------------- | ----------- | -------- | --------- | ---------- | ---------- |
| ai.cs.sp2cart | 0.6371s     | 0.1066s | 0.0605s | 5.97x     | 10.53x    |

(3) Package photutils provides an API to find the best fit of first and
second harmonic functions to a set of (angle, intensity) pairs:
Array size = 1000, Command = "%timeit fit_first_and_second_harmonics(E, data)"

| Function Name | NumPy 1.16 | AVX2   | AVX512  | AVX2 speed up   | AVX512  speed up |
| ------------- | ----------- | -------- | --------- | ---------- | ---------- |
| fit_first_and_second_harmonics | 1.598s      | 0.8709s | 0.7761s | 1.83x     | 2.05x   |