@mattip 2019-04-23
Sorry for not being clear. I meant that I would expect that a performance PR would come with changes in benchmarks in `numpy.benchmarks`, run by 
```
python runtests.py --bench
```