@r-devulap 2019-04-23
I will take a look at this today and let you know if I am able to run the benchmarks for float32.