@r-devulap 2019-04-23
I ran the ufunc benchmarks `python  runtests.py --bench ufunc` and saw this: 

| Function | Before this patch | With this patch |
| -- | -- | -- |
| sin | 7.51±0.01ms | 7.15±0ms |
| cos | 7.73±0.02ms | 7.36±0ms | 

But the thing is, the benchmarks for ufunc seems to reporting just one metric (sum of the total time, I assume) for a whole bunch of dtypes (int16, int32, float16, float32, float128, complex64, complex256), whereas my patch only improves performance for float32. So I am not sure if these numbers are representative of performance of this patch.

I added my own benchmarks which measures computing sine and cosine for just float32: 

```
class Trignometry(Benchmark):                                                   
    def setup(self):                                                            
        self.f = np.ones(10000, dtype=np.float32)                           
                                                                                
    def time_sine(self):                                                        
        np.sin(self.f)                                                          
                                                                                
    def time_cosine(self):                                                      
        np.cos(self.f)  
```

and this is what I measured: 

| Function | Before this patch | With this patch |
| -- | -- | -- |
| sin | 117±0.04us | 12.2±1us |
| cos | 115±0.02us | 12.9±1us |