@mattip 2019-04-30
This makes `_multiarray_umath.so` on Ubuntu 18.04 with gcc-8 grow from 17271072 to 17299296, which is 27.5 kB or 0.16%.