@r-devulap 2019-05-01
Made some edits to handle NAN and added some tests as well.