@r-devulap 2019-05-07
Added a commit which fixes the bug reported in https://github.com/numpy/numpy/issues/13493 for sin and cos.