@r-devulap 2019-05-23
Updated with following fixes: 
- using AVX for strided arrays (fixes the same issue reported for exp/log)
- sin/cos will require an fma instead of just avx2. Otherwise the results are in-consistent across avx2 and avx512 implementations. 
- Rounded up the max tolerable ulp error to nearest int (the `assert_array_max_ulp `returns an int)