@r-devulap 2019-05-23
yeah, thanks :) It was in cpuid.c, just fixed it.