@mattip 2019-06-25
I am OK with this, it does not add much bloat and gives a considerable performance boost. @juliantaylor any thoughts?