@mattip 2019-07-14
I am a bit torn on this. I would like to merge it, but would like to have gh-13515 in place (tests for floating point accuracy) first.