@r-devulap 2019-07-16
> I am a bit torn on this. I would like to merge it, but would like to have [gh-13515](https://github.com/numpy/numpy/issues/13515) in place (tests for floating point accuracy) first.

sounds good, I will get a patch going and submit a separate PR for this. As pointed out by @juliantaylor, glibc has nice list of floating point numbers to test each function (example: [here](https://sourceware.org/git/?p=glibc.git;a=blob;f=math/auto-libm-test-out-cos;h=072e04f157123e5439995a35a4037f67ded2af1f;hb=HEAD)). I can get started with this list first.