@r-devulap 2019-07-19
@mattip PR #14048 has the test suite covering float32 exp, log, sin and cos. Let me know if that looks good.