@r-devulap 2019-08-17
@mattip ping, how do you want to proceed with PR? Is there a way to enable the tests only for AVX machines?