@r-devulap 2019-08-17
The test `test_validate_transcendentals` in the file `numpy/core/tests/test_umath_accuracy.py` is currently disabled completely in the CI. The reason is that it is different platforms/compilers seem to generate different outputs for the same functions and the accuracy varies significantly. I would like to know if there is way to enable this test in the CI only for platforms that NumPy utilizes the AVX version of sin/cos/exp/log and ignore the test on other platforms.