@mattip 2019-08-17
I guess you could add a function to `numpy/core/src/umath/_umath_tests.c` which would export a C-level detection of AVX, then call it as part of the setup for those tests.

Would you like to do that as part of this PR or in a separate, follow-on one?